requirements.txt
alpaca-py>=0.8.0
pandas>=1.5.0
numpy>=1.21.0
python-dotenv>=0.19.0
backtrader>=1.9.76.123
matplotlib>=3.5.0
seaborn>=0.12.0
tabulate>=0.9.0